Update Constraints.adoc
From what I can find, `size` can no longer be used to constrain a numeric value, which has been deprecated. It can only constrain text fields or size of a Collection. (So, I removed the phrase "or number".)

I also changed the example to zipcode which is a string that should be limited to 5-10 characters, instead of "children" which ... I don't know what they were limiting to 5-15 there, but it didn't make much sense.